requirements.txt
pandas
numpy
requests
scikit-learn
joblib
apache-airflow>=2.0,<3
